+++ R/esquisserUI.R
@@ -1,59 +1,155 @@
-#' Ui for addin esquisser
+#' @title Esquisse Shiny module
+#' 
+#' @description Launch \code{esquisse} in a classic Shiny app.
 #'
+#' @param id Module's id.
+#' @param header Logical. Display or not \code{esquisse} header.
+#' @param choose_data Logical. Display or not the button to choose data.
+#' 
+#' @note For the module to display correctly, it is necessary to place it in a container with a fixed height.
 #'
-#' @return a \code{shiny::\link[shiny]{tagList}} containing UI elements
-#' @noRd
+#' @export
+#' 
+#' @name esquisse-module
 #'
-#' @importFrom htmltools tags tagList
-#' @importFrom shiny plotOutput icon actionButton
+#' @importFrom htmltools tags tagList singleton
+#' @importFrom shiny plotOutput icon actionButton NS
 #' @importFrom miniUI miniTitleBarButton miniPage
 #'
-esquisserUI <- function() {
+#' @examples 
+#' 
+#' if (interactive()) {
+#' 
+#' 
+#' ### Part of a Shiny app ###
+#' 
+#' library(shiny)
+#' library(esquisse)
+#' 
+#' ui <- fluidPage(
+#'   tags$h1("Use esquisse as a Shiny module"),
+#'   radioButtons(
+#'     inputId = "data", 
+#'     label = "Data to use:", 
+#'     choices = c("iris", "mtcars"),
+#'     inline = TRUE
+#'   ),
+#'   tags$div(
+#'     style = "height: 700px;", # needs to be in fixed height container
+#'     esquisserUI(
+#'       id = "esquisse", 
+#'       header = FALSE, # dont display gadget title
+#'       choose_data = FALSE # dont display button to change data
+#'     )
+#'   )
+#' )
+#' 
+#' server <- function(input, output, session) {
+#'   
+#'   data_r <- reactiveValues(data = iris, name = "iris")
+#'   
+#'   observeEvent(input$data, {
+#'     if (input$data == "iris") {
+#'       data_r$data <- iris
+#'       data_r$name <- "iris"
+#'     } else {
+#'       data_r$data <- mtcars
+#'       data_r$name <- "mtcars"
+#'     }
+#'   })
+#'   
+#'   callModule(module = esquisserServer, id = "esquisse", data = data_r)
+#'   
+#' }
+#' 
+#' shinyApp(ui, server)
+#' 
+#' 
+#' 
+#' ### Whole Shiny app ###
+#' 
+#' library(shiny)
+#' library(esquisse)
+#' 
+#' 
+#' # Load some datasets in app environment
+#' my_data <- data.frame(
+#'   var1 = rnorm(100),
+#'   var2 = sample(letters[1:5], 100, TRUE)
+#' )
+#' 
+#' 
+#' 
+#' ui <- fluidPage(
+#'   tags$div( # needs to be in fixed height container
+#'     style = "position: fixed; top: 0; bottom: 0; right: 0; left: 0;", 
+#'     esquisserUI(id = "esquisse")
+#'   )
+#' )
+#' 
+#' server <- function(input, output, session) {
+#'   
+#'   callModule(module = esquisserServer, id = "esquisse")
+#'   
+#' }
+#' 
+#' shinyApp(ui, server)
+#' 
+#' }
+#' 
+esquisserUI <- function(id, header = TRUE, choose_data = TRUE) {
+  
+  ns <- NS(id)
+  
+  
+  box_title <- tags$div(
+    class="gadget-title dreamrs-title-box",
+    tags$h1(shiny::icon("wrench"), "ggplot2 builder", class = "dreamrs-title"),
+    tags$div(
+      class = "pull-right",
+      miniTitleBarButton(inputId = ns("close"), label = "Close")
+    )
+  )
+    
 
   ### addin
   miniPage(
 
     # style sheet
-    tags$link(rel="stylesheet", type="text/css", href="esquisse/styles.css"),
-    tags$script(src = "esquisse/clipboard.min.js"),
+    singleton(x = tagList(
+      tags$link(rel="stylesheet", type="text/css", href="esquisse/styles.css"),
+      tags$script(src = "esquisse/clipboard/clipboard.min.js")
+    )),
 
-    # title
-    tags$div(
-      class="gadget-title dreamrs-title-box",
-      tags$h1(shiny::icon("wrench"), "ggplot2 builder", class = "dreamrs-title"),
-      tags$div(
-        class = "pull-right",
-        miniTitleBarButton(inputId = "close", label = "Close")
-      )
-    ),
+    if (isTRUE(header)) box_title,
     # page
     layoutAddin(
       top_left = htmltools::tagList(
         htmltools::tags$div(
           style = "padding: 10px;",
           imageButtonUI(
-            id = "geom",
+            id = ns("geom"),
             imgs = geom_icon_input(), 
             width = "240px"
           ),
-          htmltools::tags$br(),
-          chooseDataUI(id = "choose-data")
+          # htmltools::tags$br(),
+          if (isTRUE(choose_data)) chooseDataUI(id = ns("choose-data"))
         )
       ),
       top_right = dragulaInput(
-        inputId = "dragvars", sourceLabel = "Variables", 
-        targetsLabels = c("X", "Y", "Fill", "Color", "Size"), 
-        targetsIds = c("xvar", "yvar", "fill", "color", "size"),
+        inputId = ns("dragvars"), sourceLabel = "Variables", 
+        targetsLabels = c("X", "Y", "Fill", "Color", "Size", "Facet"), 
+        targetsIds = c("xvar", "yvar", "fill", "color", "size", "facet"),
         choices = "", badge = FALSE, width = "100%", height = "100%",
         replace = TRUE
       ),
       main = htmltools::tags$div(
         style = "margin-top: 10px; padding-bottom: 25px; height: 100%;",
-        shiny::plotOutput(outputId = "plooooooot", width = "100%", height = "100%")
+        shiny::plotOutput(outputId = ns("plooooooot"), width = "100%", height = "100%")
       )
     ),
 
-    chartControlsUI(id = "controls")
+    chartControlsUI(id = ns("controls"))
   )
 
 }