+++ R/module-code.R
@@ -16,12 +16,12 @@ moduleCodeUI <- function(id) {
   htmltools::tagList(
     tags$button(
       class = "btn btn-default btn-xs pull-right btn-copy-code",
-      "Copy to clipboard", `data-clipboard-target` = "#codeggplot"# onclick = "ClipBoard()"
+      "Copy to clipboard", `data-clipboard-target` = paste0("#", ns("codeggplot"))
     ), htmltools::tags$script("new Clipboard('.btn-copy-code');"),
     htmltools::tags$br(),
     htmltools::tags$b("Code:"),
     shiny::uiOutput(outputId = ns("code")),
-    htmltools::tags$textarea(id = "holderCode", style = "display: none;"),
+    htmltools::tags$textarea(id = ns("holderCode"), style = "display: none;"),
     shiny::actionLink(
       inputId = ns("insert_code"),
       label = "Insert code in script",
@@ -52,22 +52,23 @@ moduleCodeUI <- function(id) {
 moduleCodeServer <- function(input, output, session, varSelected, dataChart, paramsChart, geomSelected) {
 
   ns <- session$ns
+  # print(ns("code"))
 
   codegg <- shiny::reactive({
     code_geom <- guess_geom(
-      xtype = if (!is.null(varSelected$xvar)) col_type(dataChart$data[[varSelected$xvar]]),
-      ytype = if (!is.null(varSelected$yvar)) col_type(dataChart$data[[varSelected$yvar]]),
+      xtype = if (!is.null(varSelected$x)) col_type(dataChart$data[[varSelected$x]]),
+      ytype = if (!is.null(varSelected$y)) col_type(dataChart$data[[varSelected$y]]),
       type = geomSelected$x, sfobj = inherits(dataChart$data, what = "sf")
     )
     code_aes <- guess_aes(
-      x = varSelected$xvar,
-      y = varSelected$yvar,
+      x = varSelected$x,
+      y = varSelected$y,
       fill = varSelected$fill,
       color = varSelected$color,
       size = varSelected$size,
       geom = code_geom,
-      xtype = if (!is.null(varSelected$xvar)) col_type(dataChart$data[[varSelected$xvar]]),
-      ytype = if (!is.null(varSelected$yvar)) col_type(dataChart$data[[varSelected$yvar]])
+      xtype = if (!is.null(varSelected$x)) col_type(dataChart$data[[varSelected$x]]),
+      ytype = if (!is.null(varSelected$y)) col_type(dataChart$data[[varSelected$y]])
     )
     code_aes <- lapply(
       X = code_aes,
@@ -132,6 +133,7 @@ moduleCodeServer <- function(input, output, session, varSelected, dataChart, par
       geom = code_geom,
       scale = code_scale,
       args_geom = args_geom,
+      facet = varSelected$facet,
       theme = params_chart$theme, coord = coord,
       labs = params_chart[c("title", "x", "y", "caption", "subtitle")],
       params = params_chart
@@ -151,11 +153,7 @@ moduleCodeServer <- function(input, output, session, varSelected, dataChart, par
 
   output$code <- shiny::renderUI({
     htmltools::tagList(
-      rCodeContainer(id = "codeggplot", codegg())#,
-      # htmltools::tags$button(
-      #   class="btn btn-clipboard", "Copy", `data-clipboard-target`="#code_ggplot"
-      # ),
-      # htmltools::tags$script("new Clipboard('btn-clipboard');")
+      rCodeContainer(id = ns("codeggplot"), codegg())
     )
   })