+++ R/module-imageButton.R
@@ -24,7 +24,7 @@ imageButtonUI <- function(id, imgs = list(), selected = 1, up = FALSE, width = N
     actionButton(
       inputId = inputId,
       label = tagList(
-        tags$img(src = img, width = 80, height = 80),
+        tags$img(src = img, width = 70, height = 70),
         tags$br(), label
       ),
       style = "border: none;"
@@ -41,9 +41,9 @@ imageButtonUI <- function(id, imgs = list(), selected = 1, up = FALSE, width = N
     class = "dropdown-toggle", 
     `data-toggle` = "dropdown"
   )
-  dropTag <- tags$ul(
-    class = "dropdown-menu",
-    style = "padding: 5px;",
+  dropTag <- tags$div(
+    class = "dropdown-menu pre-scrollable",
+    style = "padding: 5px; max-height: 80vh;",
     style = if (!is.null(width))
       paste0("width: ", validateCssUnit(width), ";"),
     lapply(
@@ -59,9 +59,10 @@ imageButtonUI <- function(id, imgs = list(), selected = 1, up = FALSE, width = N
   )
 
   tagList(
-    toggleInputUi(),
+    useShinyUtils(),
     tags$div(
       class = ifelse(up, "dropup", "dropdown"),
+      style = "margin-bottom: 1vh;",
       btn, dropTag
     ),
     tags$script(
@@ -107,9 +108,9 @@ imageButtonServer <- function(input, output, session, default = NULL, img_ref = 
     if (!is.null(enabled$x)) {
       for (i in names(img_ref)) {
         if (i %in% enabled$x) {
-          toggleInputServer(session = session, inputId = ns(i), enable = TRUE)
+          toggleInput(session = session, inputId = ns(i), enable = TRUE)
         } else {
-          toggleInputServer(session = session, inputId = ns(i), enable = FALSE)
+          toggleInput(session = session, inputId = ns(i), enable = FALSE)
         }
       }
     }